Comments

+++ app/Http/Controllers/NacteResultController.php
@@ -36,9 +36,11 @@ class NacteResultController extends Controller
     public function destroy(Request $request)
     {
     	$detail = NacteResultDetail::find($request->get('nacte_result_detail_id'));
+        if($detail->verified != 1){
     	NacteResult::where('nacte_result_detail_id',$request->get('nacte_result_detail_id'))->delete();
     	// $detail->results->delete();
     	$detail->delete();
+        }
 	    return redirect()->back()->with('message','NACTE results declined successfully');
     }
